Consolidate the two example programs ...
... and make the simple one even simpler.

+++ example-project-with-subdir/src/capi.rs
@@ -1,3 +1,5 @@
+use libc::size_t;
+
 use crate::add_two;
 
 // NB: The documentation comments from this file will be available
@@ -5,6 +7,6 @@ use crate::add_two;
 
 /// Adds two to the given value and returns the result.
 #[no_mangle]
-pub extern "C" fn example_project_with_subdir_add_two(value: u32) -> u32 {
+pub extern "C" fn example_project_with_subdir_add_two(value: size_t) -> size_t {
     add_two(value)
 }
