+++ example-project-with-subdir/src/lib.rs
@@ -2,14 +2,13 @@
 //!
 //! [cargo-c]: https://crates.io/crates/cargo-c
 
-#![warn(rust_2018_idioms)]
 #![deny(missing_docs)]
 
 #[cfg(feature = "capi")]
 mod capi;
 
 /// Adds two to the given value and returns the result.
-pub fn add_two(value: u32) -> u32 {
+pub fn add_two(value: usize) -> usize {
     value + 2
 }
